requirements-dev.txt
aiosmtpd~=1.4.6
asgiref~=3.11.1
boto3~=1.34.162
botocore~=1.34.162
cachetools>=5.3.0
celery~=5.6.3
django~=4.1.13
djangorestframework~=3.15.1
httpx~=0.28.1
imageio~=2.33.1
locust>=2.29.0
matplotlib~=3.8.4
numpy~=1.26.4
pandas~=2.2.3
pillow~=10.3.0
py2neo~=2021.2.4
pyjwt~=2.12.1
pymongo~=4.16.0
pymysql~=1.1.2
pytest>=8.0.0
pytest-benchmark>=4.0.0
pytest-cov>=4.0.0
readline~=8.3
zlib~=1.2.13
